Scenario 3: use article titles as keys
Turns out the values that I’ve been using are not actually the article
IDs but rather the IDs of the events. That’s why the total count was
always 1 — doh!

Anyway, it’s a little odd to use the page title — a string — rather than
a page ID — a number, presumably — but it appears these events don’t
include the page IDs‽

https://www.mediawiki.org/wiki/Manual:RCFeed

¯\_(ツ)_/¯